Squashed commit of the following:
commit c3606eb457b02d915d87b329208cf52ab38a08e0
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Tue Mar 18 10:50:13 2025 +0100

    Reapply "fix: change visibility of header links"

    This reverts commit 42a7ae3dc6717d81902844abead0421071694531.

commit e4a522e62d32f6bbfc1ed0c49571515b701c33ad
Merge: 7ff5856a 2ff580f1
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Tue Mar 18 10:36:32 2025 +0100

    Merge branch 'main' into mobile-header-improvements

commit 7ff5856a980a2eacf2290be2f77160e841dc44b3
Merge: 7c47e38a 578798f8
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Tue Mar 18 10:35:50 2025 +0100

    Merge branch 'mobile-header-improvements' of github.com:smg-automotive/components-pkg into mobile-header-improvements

commit 7c47e38a4894bf5a7c92fbe693632fcf3bf9f3a7
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Tue Mar 18 10:29:57 2025 +0100

    fix: the switch to mobile design should happen on 'sm' not 'xs'

commit 42a7ae3dc6717d81902844abead0421071694531
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Tue Mar 18 10:20:13 2025 +0100

    Revert "fix: change visibility of header links"

    This reverts commit 5b864a7b08f04887f7751f36903f296a9b20ceb6.

commit 2ff580f1c4abbe31c528eaaad8db58dcddcafa5b
Author: Nikola-Suvajcevic <nikola.suvajcevic@swissmarketplace.group>
Date:   Tue Mar 18 09:48:09 2025 +0100

    fix: update security advice links in footer (#1345)

commit 541c86fced6ced39aaabfc6511d8c2fc29ae4b59
Author: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Date:   Tue Mar 18 09:03:22 2025 +0100

    fix(deps): update dependency @chakra-ui/react to v2.10.7 (#1344)


commit 578798f82d6c99d1620cef599e40ccc1bdebef1e
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 17:53:40 2025 +0100

    chore: initialise the mock function


commit 69283886ed30137734a936fdd2daf334dedccedf
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 17:53:16 2025 +0100

    chore: remove trailing space


commit 1f2935b562b5b7c47a182ddd47ef8763ee57631f
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 17:25:32 2025 +0100

    chore: re-format

commit 00a2ad474220378bc5472e7ee35e4f2aa6196c6c
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 17:23:50 2025 +0100

    fix: hide user items if the user is logged out

commit ed960ad38f45b62d9e3ca39cf7a2581a5bebf8bc
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 17:01:30 2025 +0100

    fix: do not shrink the navigation avatar

commit 1497c741c9b4847e5dbc7d2b01471f08bc6f5f32
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 16:51:25 2025 +0100

    feat: add login toggle

commit 416a3eb7da5f3f6c03f17176bd1083d299beddd6
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 16:45:56 2025 +0100

    feat: logout icon

commit 195e409f61ee59c2d955351541123d502661dc5f
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 16:34:43 2025 +0100

    feat: language selection in combined menu

commit d87d19b3f775b309cbcba2b6c4ef3b7ccff68525
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 16:34:25 2025 +0100

    chore: extract language replacing

commit d202f029e26d4a5cbee2400f115b2bfbd2884f3f
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 16:34:03 2025 +0100

    feat: allow passing title directly to link configuration

commit 40cb2337b190eb771bf5db1f52027563aa89efff
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:56:00 2025 +0100

    chore: split ant extact header menu items configuration

commit c4ff7537465d2cfdaab66c17c32d43ea203e5ae1
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:15:14 2025 +0100

    feat: combined menu toggle

commit c33de4dcc6ed54478ee8e2cf19b8d0afcf44c607
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:09:20 2025 +0100

    fix: render user info in combined menu

commit 4c3719b7211f711e90d65afa8b887506e75a7775
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:08:53 2025 +0100

    feat: hamburger menu item

commit 947cca28a9a4a5363441c4363bf3071fe5a56ad6
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:08:39 2025 +0100

    feat: combined menu configuration

commit 5bcd2e2db70ca81b0e461ef65423212255df1b52
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:06:30 2025 +0100

    fix: ensure drawer is scrollable on mobile

commit 5b864a7b08f04887f7751f36903f296a9b20ceb6
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:05:58 2025 +0100

    fix: change visibility of header links

commit 014099883db91d3933abdcebe356f0d6da7a4960
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:05:06 2025 +0100

    fix: use placement in the language menu

commit 4143d9b1abbbae658fa62240762f94efb492503f
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:04:11 2025 +0100

    feat: hide search menu on mobile

commit aa3f84e3754682eab76bd9aed89b391dccbb0bfd
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:02:58 2025 +0100

    chore: simplify color logic

commit 7dd664e06b5d936844e0f86151dc8264c291cd4a
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:02:36 2025 +0100

    fix: hide user menu on mobile

commit 5a0748c564a533393a2ddcf1bb9867acc1910671
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:01:48 2025 +0100

    fix: exhaustive deps

commit 289515058412c9fd342cece1f84d48213d499423
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 15:00:13 2025 +0100

    chore: keep old entitlements

commit ea600bfc0dcff5d58aa78cc3f45d892cb2a24cc4
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:59:27 2025 +0100

    chore: restructure header test to use fixtures

commit eb6f43ab47f3cf6f89a3bf11c3f6ef5c29491b1d
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:55:24 2025 +0100

    chore: extract user to fixtures

commit 59f19f48aff41de8c4e1c321708abc135eae58a3
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:53:30 2025 +0100

    fix: comparison item spacing

commit ac25da6d021b6c1889cd8903f8d590f69bd9e627
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:52:54 2025 +0100

    chore: fix spacing messed up by git

commit 1f25da86a9d2b398758fe1df1a11877e2189d86e
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:52:04 2025 +0100

    fix: missed imports and types

commit b4d2f20641460af426eef0a3a6051572997ac750
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:51:16 2025 +0100

    fix: highlight Menu when open

commit ca740938a8edfaf02c9253762dc4ab1e054a05a3
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:50:37 2025 +0100

    fix: animate the icon when Menu is open

commit 67e6fdf32d54b9f948d61a34f67c3bbdbc68fb8b
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:49:48 2025 +0100

    feat: support icon, spacing and placement props for Menu

commit ae53ba7c28a88b4019b1c93decfb5b037b8ce074
Author: Krzysztof Sakwerda <ksakwerda@gmail.com>
Date:   Fri Mar 14 14:45:02 2025 +0100

    fix: add transition to the drawer indicator
