+++ package.json
@@ -121,7 +121,7 @@
   },
   "dependencies": {
     "@chakra-ui/anatomy": "2.3.4",
-    "@chakra-ui/react": "2.10.6",
+    "@chakra-ui/react": "2.10.7",
     "@chakra-ui/styled-system": "2.12.0",
     "@chakra-ui/system": "2.6.2",
     "@chakra-ui/theme-tools": "2.2.6",
