+++ src/components/icons/index.ts
@@ -150,3 +150,5 @@ export { TriangleIcon } from './TriangleIcon';
 export { SwissFlagIcon } from './SwissFlagIcon';
 export { GlobeIcon } from './GlobeIcon';
 export { GarageIcon } from './GarageIcon';
+export { HamburgerMenuIcon } from './HamburgerMenuIcon';
+export { LogoutIcon } from './LogoutIcon';
