+++ src/components/navigation/header/drawer/index.tsx
@@ -9,8 +9,10 @@ import DrawerContent from 'src/components/drawer/DrawerContent';
 import Drawer from 'src/components/drawer';
 
 import { Drawer as useNavigationDrawerType } from '../hooks/useNavigationDrawer';
+import { DrawerNode } from '../config/DrawerNodeItems';
 import DrawerUserInfo from './UserInfo';
 import { DrawerMenu } from './DrawerMenu';
+import DrawerLoginToggle from './DrawerLoginToggle';
 
 interface NavigationDrawerProps {
   user: EnrichedSessionUser | null;
@@ -18,6 +20,8 @@ interface NavigationDrawerProps {
   isOpen: boolean;
   onClose: () => void;
   menuHeight: string;
+  onLogin: () => void;
+  onLogout: () => void;
 }
 
 export const NavigationDrawer: FC<NavigationDrawerProps> = ({
@@ -26,11 +30,13 @@ export const NavigationDrawer: FC<NavigationDrawerProps> = ({
   onClose,
   menuHeight,
   user,
+  onLogin,
+  onLogout,
 }) => {
   return (
     <Drawer isOpen={isOpen} placement="top" onClose={onClose}>
       <DrawerOverlay />
-      <DrawerContent marginTop={menuHeight}>
+      <DrawerContent marginTop={menuHeight} overflow="scroll" maxH="100%">
         <DrawerBody
           data-testid="drawer-body"
           py="lg"
@@ -44,10 +50,21 @@ export const NavigationDrawer: FC<NavigationDrawerProps> = ({
             templateColumns={{ '2xs': '1fr', md: 'repeat(5, 1fr)' }}
             gridGap={{ md: '3xl' }}
           >
-            {drawer?.current === 'user' ? <DrawerUserInfo user={user} /> : null}
+            {[DrawerNode.User, DrawerNode.Combined].includes(
+              drawer?.current as DrawerNode,
+            ) ? (
+              <DrawerUserInfo user={user} />
+            ) : null}
             {drawer?.nodes.map((node, index) => (
               <DrawerMenu key={`node-${index}`} node={node} />
             ))}
+            {drawer?.current === DrawerNode.Combined ? (
+              <DrawerLoginToggle
+                user={user}
+                onLogin={onLogin}
+                onLogout={onLogout}
+              />
+            ) : null}
           </Grid>
         </DrawerBody>
       </DrawerContent>
